+++ packages/backend/src/core/activitypub/models/ApNoteService.ts
@@ -22,7 +22,6 @@ import { IdService } from '@/core/IdService.js';
 import { PollService } from '@/core/PollService.js';
 import { StatusError } from '@/misc/status-error.js';
 import { UtilityService } from '@/core/UtilityService.js';
-import { bindThis } from '@/decorators.js';
 import { checkHttps } from '@/misc/check-https.js';
 import { IdentifiableError } from '@/misc/identifiable-error.js';
 import { getOneApId, getApId, getOneApHrefNullable, validPost, isEmoji, getApType } from '../type.js';
@@ -78,6 +77,12 @@ export class ApNoteService implements OnModuleInit {
 		private apLoggerService: ApLoggerService,
 	) {
 		this.logger = this.apLoggerService.logger;
+		this.validateNote = this.validateNote.bind(this);
+		this.fetchNote = this.fetchNote.bind(this);
+		this.createNote = this.createNote.bind(this);
+		this.resolveNote = this.resolveNote.bind(this);
+		this.parseAndValidateGeoShare = this.parseAndValidateGeoShare.bind(this);
+		this.extractEmojis = this.extractEmojis.bind(this);
 	}
 
 	onModuleInit(): void {
@@ -90,7 +95,6 @@ export class ApNoteService implements OnModuleInit {
 		this.apDbResolverService = this.moduleRef.get('ApDbResolverService');
 	}
 
-	@bindThis
 	public validateNote(object: IObject, uri: string, actor?: MiRemoteUser): Error | null {
 		const expectHost = this.utilityService.extractDbHost(uri);
 		const apType = getApType(object);
@@ -128,7 +132,6 @@ export class ApNoteService implements OnModuleInit {
 	 *
 	 * Misskeyに対象のNoteが登録されていればそれを返します。
 	 */
-	@bindThis
 	public async fetchNote(object: string | IObject): Promise<MiNote | null> {
 		return await this.apDbResolverService.getNoteFromApId(object);
 	}
@@ -136,7 +139,6 @@ export class ApNoteService implements OnModuleInit {
 	/**
 	 * Noteを作成します。
 	 */
-	@bindThis
 	public async createNote(value: string | IObject, actor?: MiRemoteUser, resolver?: ApResolverLike, silent = false): Promise<MiNote | null> {
 		// eslint-disable-next-line no-param-reassign
 		if (resolver == null) resolver = await this.apResolverService.createResolver();
@@ -368,7 +370,6 @@ export class ApNoteService implements OnModuleInit {
 	 * Misskeyに対象のNoteが登録されていればそれを返し、そうでなければ
 	 * リモートサーバーからフェッチしてMisskeyに登録しそれを返します。
 	 */
-	@bindThis
 	public async resolveNote(value: string | IObject, options: { sentFrom?: URL, resolver?: ApResolverLike } = {}): Promise<MiNote | null> {
 		const uri = getApId(value);
 
@@ -398,7 +399,6 @@ export class ApNoteService implements OnModuleInit {
 		}
 	}
 
-	@bindThis
 	public parseAndValidateGeoShare(geoShareData: any): string | null {
 		if (!geoShareData) return null;
 
@@ -451,7 +451,6 @@ export class ApNoteService implements OnModuleInit {
 		}
 	}
 
-	@bindThis
 	public async extractEmojis(tags: IObject | IObject[], host: string): Promise<MiEmoji[]> {
 		// eslint-disable-next-line no-param-reassign
 		host = this.utilityService.toPuny(host);